Code review: Separate `db-types.ts` from `types.ts`
