+++ src/App.tsx
@@ -3,7 +3,8 @@ import * as randomize from "randomatic";
 import { CodapHelper as Codap, ISaveState} from "./lib/codap-helper";
 import codapInterface, { ClientNotification } from "./lib/CodapInterface";
 import { DB } from "./lib/db";
-import { DataContext, CodapItem, DBSharedTable } from "./lib/types";
+import { DBSharedTable } from "./lib/db-types";
+import { DataContext, CodapItem } from "./lib/types";
 const pkg = require("../package.json");
 import "./App.css";
 import pDebounce from "p-debounce";
