+++ src/lib/types.ts
@@ -45,23 +45,3 @@ export interface CodapItem {
   id: string;
   values: CodapItemValues;
 }
-
-export interface DBUsers {
-  [user: string]: string;
-}
-
-export interface DBItemData {
-  items: { [id: string]: CodapItemValues };
-  order: string[];
-}
-
-export interface DBItemDataMap {
-  [user: string]: DBItemData;
-}
-
-export interface DBSharedTable {
-  connectedUsers: DBUsers;
-  allUsers: DBUsers;
-  dataContext: DataContext;
-  itemData?: DBItemDataMap;
-}
